nanovg.rb: Now compatible with the latest APIs (NanoVG Commit f4069e6a1ad5da430fb0a9c57476d5ddc2ff89b2 https://github.com/memononen/nanovg/commit/f4069e6a1ad5da430fb0a9c57476d5ddc2ff89b2 ).

+++ nanovg.rb
@@ -138,7 +138,7 @@ module NanoVG
     #
     # Common API
     #
-    attach_function :nvgBeginFrame, :nvgBeginFrame, [:pointer, :int32, :int32, :float], :void
+    attach_function :nvgBeginFrame, :nvgBeginFrame, [:pointer, :float, :float, :float], :void
     attach_function :nvgCancelFrame, :nvgCancelFrame, [:pointer], :void
     attach_function :nvgEndFrame, :nvgEndFrame, [:pointer], :void
